move isAuthorized into AuthProvider, and remove HOC because of uneeded complexity
